.claude/worktrees/plugin-edition/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "assistonauts"
version = "0.1.0"
description = "LLM-powered knowledge base framework with specialized AI agents"
requires-python = ">=3.11"
dependencies = [
    "litellm",
    "pyyaml",
    "click",
    "rich",
    "markitdown",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
]

[project.scripts]
assistonauts = "assistonauts.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/assistonauts"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
src = ["src"]
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "RUF",   # ruff-specific
]

[tool.ruff.lint.isort]
known-first-party = ["assistonauts"]
